requirements.txt
torch
streamlit
Pillow
numpy
